Updated docs and cleaned up some tests
