+++ tests/test_simple_examples.py
@@ -1,12 +1,11 @@
 import os
 import yaml
 from copy import deepcopy
-from tests.helpers import set_up_s3
-from contextlib import contextmanager
-import io
-import boto3
 from pyarrow import fs
-from dataengineeringutils3.s3 import s3_path_to_bucket_key
+from tests.helpers import (
+    set_up_s3,
+    mock_get_file,
+)
 
 simple_yaml_config = """
 land-base-path: s3://land/
@@ -32,30 +31,14 @@ tables:
 """
 
 
-class MockS3FilesystemReadInputStream:
-    @staticmethod
-    @contextmanager
-    def open_input_stream(s3_file_path_in: str) -> io.BytesIO:
-        s3_resource = boto3.resource("s3")
-        bucket, key = s3_path_to_bucket_key(s3_file_path_in)
-        obj_bytes = s3_resource.Object(bucket, key).get()["Body"].read()
-        obj_io_bytes = io.BytesIO(obj_bytes)
-        try:
-            yield obj_io_bytes
-        finally:
-            obj_io_bytes.close()
-
-
 def test_validation_single_worker(s3, monkeypatch):
     """
     Simple example on how to run DL for a single worker.
 
     [init] -> [worker]x1 -> [closedown]
     """
-    
+
     # Need to mock S3 read for pyarrow (only for testing)
-    def mock_get_file(*args, **kwargs):
-        return MockS3FilesystemReadInputStream()
     monkeypatch.setattr(fs, "S3FileSystem", mock_get_file)
 
     from data_linter import validation
@@ -85,9 +68,6 @@ def test_validation_multiple_workers(s3, monkeypatch):
 
     [init] -> [worker]x4 -> [closedown]
     """
-
-    def mock_get_file(*args, **kwargs):
-        return MockS3FilesystemReadInputStream()
 
     monkeypatch.setattr(fs, "S3FileSystem", mock_get_file)
 
@@ -125,9 +105,6 @@ def test_validation_multiple_workers_no_init(s3, monkeypatch):
 
     [worker]x2 -> [closedown]
     """
-
-    def mock_get_file(*args, **kwargs):
-        return MockS3FilesystemReadInputStream()
 
     monkeypatch.setattr(fs, "S3FileSystem", mock_get_file)
